+++ test/integration/test_call_variant_peptides.py
@@ -1048,3 +1048,16 @@ class TestCallVariantPeptides(TestCaseIntegration):
         expected = self.data_dir/'fuzz/40/brute_force.txt'
         reference = self.data_dir/'fuzz/40'
         self.default_test_case(gvf, reference, expected)
+
+    def test_call_variant_peptide_case70(self):
+        """ In peptide graph for noncoding transcripts, all possible ORFs are
+        kept for each node, but only one cursor (inbond node to outbound node)
+        is used per node. This causes a problem for circRNA because the cleavage
+        gain variant maybe added to wrong orf. """
+        gvf = [
+            self.data_dir/'fuzz/41/fake_variants.gvf',
+            self.data_dir/'fuzz/41/fake_circ_rna.gvf'
+        ]
+        expected = self.data_dir/'fuzz/41/brute_force.txt'
+        reference = self.data_dir/'fuzz/41'
+        self.default_test_case(gvf, reference, expected)