+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Fixed that hybrid node not identified if the variant is in the of an exon. #782
 
+- Fixed that in circRNA, cleavage gain from upstream node is added to the the wrong ORF. #783
+
 ## [1.2.0] - 2023-07-04
 
 ### Fixed